UPSTREAM: <carry>: update lint makefile target in openshift.mk
The upstream has renamed the makefile target from `test-lint` to `lint`,
this change reflects that in the openshift additional makefile. It also
adds some logic to skip the linting check if the linter binary is not
defined. This logic is added because the OpenShift build process
requires that all binaries be present at build time, we cannot download
a linting tool. By using the environment variable as a key to run the
linting job, we could choose to inject a linter by using a different
build image and defining `LINTER` in the container environment.