UPSTREAM: <carry>: Account for endpoint not being rerturned when building cloud configuration
This change updates AzureCloudConfigFromURL to check that the URL
(ResourceManagerEndpoint) that is used to build the config has been
returned. If it has not, it sets it using the URL provided.

This avoids the case where the endpoint is used to build config, but
does not return it's self - causing the created configuration to fail.

To make this work, we must update go.mod to replace the vendored package
with the one in the fork, and revendor. If we do not revendor, the
changes in the fork are not propagated to the /vendor directory and are
silently ignored.